backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "temple backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "PyJWT",
    "psycopg2-binary",
    "python-dotenv",
    "python-multipart",
    "boto3",
    "bcrypt",
]
